==> 0_InformacioCauca.ipynb <==
from google.colab import drive
import pandas as pd
drive.mount('/content/drive')


pip install geopandas shapely

pip install pandas requests

import geopandas as gpd
from shapely.geometry import Point

# Cargar el archivo GeoJSON
gdf = gpd.read_file("/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Datos/Veredas_de_Colombia.geojson")

# Convertir grados, minutos y segundos a decimal como se explicó anteriormente
def convertir_a_decimal(grados, minutos, segundos):
    return grados + (minutos / 60) + (segundos / 3600)

url=("/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Datos/info_vereda_1.xlsx")
df1 = pd.read_excel(io=url,sheet_name="Vereda")

url='/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Datos/BD_Editada.xlsx'
dfH=pd.read_excel(url,sheet_name='HOGAR', comment='#')
dfH.shape
dfH = pd.DataFrame(dfH)

# Convertir a coordenadas decimales
dfH['Longitud'] = dfH.apply(lambda row: -convertir_a_decimal(row['Longitud (W), Grados:'], row['Longitud (W), Minutos:'], row['Longitud (W), Segundos:']), axis=1)
dfH['Latitud'] = dfH.apply(lambda row: convertir_a_decimal(row['Latitud (N), Grados:'], row['Latitud (N), Minutos:'], row['Latitud (N), Segundos:']), axis=1)

# Crear puntos geográficos
dfH['Coordinates'] = dfH.apply(lambda row: Point(row['Longitud'], row['Latitud']), axis=1)

# Convertir DataFrame a GeoDataFrame
geo_df = gpd.GeoDataFrame(dfH, geometry='Coordinates')

# Hacer la consulta espacial
point_in_poly = gpd.sjoin(geo_df, gdf, how="left", op='intersects')

url='/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Datos/BD_Editada.xlsx'
dfE=pd.read_excel(url,sheet_name='Extras', comment='#')
dfE.shape
dfE = pd.DataFrame(dfE)

# Crear puntos geográficos
dfE['Coordinates'] = dfE.apply(lambda row: Point(row['Longitud'], row['Latitud']), axis=1)

# Convertir DataFrame a GeoDataFrame
geo_df = gpd.GeoDataFrame(dfE, geometry='Coordinates')

# Hacer la consulta espacial
point_in_polyE = gpd.sjoin(geo_df, gdf, how="left", op='intersects')

point_in_polyE = point_in_polyE[point_in_poly.columns]

# Concatenar DataFrames
df_concatenado = pd.concat([point_in_poly, point_in_polyE], ignore_index=True)

df1.rename(columns={'CÓDIGO VEREDA': 'CODIGO_VER'}, inplace=True)

df_concatenado['CODIGO_VER']=pd.to_numeric(df_concatenado['CODIGO_VER'])

df_concatenado=pd.merge(df_concatenado,df1, on='CODIGO_VER',how='left')

# Columnas a seleccionar además de COD_HOGAR
columnas_a_seleccionar = ['UP', 'UPA', 'UPNA', 'VIVIENDAS', 'HOGARES', 'PERSONAS', 'PRODUCTORES RESIDENTES', 'CodHogar']

# Crear un nuevo DataFrame con las columnas seleccionadas
nuevo_df = df_concatenado[columnas_a_seleccionar].copy()
df_concatenado=df_concatenado.drop(['UP', 'UPA', 'UPNA', 'VIVIENDAS', 'HOGARES', 'PERSONAS', 'PRODUCTORES RESIDENTES'], axis=1)

# Unimos y sumamos
resultado = nuevo_df.groupby('CodHogar').sum()

df_concatenado = df_concatenado.drop_duplicates(subset='CodHogar', keep='first')

df_concatenado=pd.merge(df_concatenado,resultado, on='CodHogar')

 df_concatenado=df_concatenado.drop(['Latitud (N), Grados:','Latitud (N), Minutos:', 'Latitud (N), Segundos:','Longitud (W), Grados:', 'Longitud (W), Minutos:','Longitud (W), Segundos:', 'Coordinates', 'index_right',], axis=1 )

# Filtrar el DataFrame para excluir filas donde la columna 'PERSONAS' es igual a 0
df_concatenado = df_concatenado[df_concatenado['PERSONAS'] != 0]

# Guardar en nuevo archivo CSV
df_concatenado.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Datos/data_concatenada.csv', index=False)
