==> 1_Creacion_de_csv.ipynb <==
pip install unidecode

from google.colab import drive
from unidecode import unidecode
from sklearn.preprocessing import KBinsDiscretizer
import numpy as np
import pandas as pd
drive.mount('/content/drive')

## Importar los Datos de la base

url='https://raw.githubusercontent.com/alejandro1232/Data_tesis/main//BD_Editada.xlsx'
df_adulto=pd.read_excel(url,sheet_name='RESULTADOS ADULTOS', comment='#')
df_adulto.shape

url='/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Datos/data_concatenada.csv'
df_hogar=pd.read_csv(url)
df_hogar.shape

data=pd.merge(df_adulto,df_hogar, left_on='g_inf_general/hogar', right_on='CodHogar')
data.shape

data['Vereda'] = data['Vereda:'] + ',' + data['Municipio:']

Variable='Vereda'

data=data.apply(lambda x: x.astype(str).str.strip().str.upper().str.replace(' ', ''))
data[Variable] = data[Variable].apply(lambda x: unidecode(x))
total=data[[Variable,'g_resultado/resultado']]
total=total.groupby(Variable).count()
total=total.rename(columns={"g_resultado/resultado": "Total"})

total.shape

positivos=data[data['g_resultado/resultado']=='POSITIVO']
positivos=positivos[[Variable,'g_resultado/resultado']]
positivos=positivos.groupby([Variable]).count()
positivos=positivos.rename(columns={"g_resultado/resultado": "Positivos"})

## Calculo de la Prevalencia

## Caracteristicas del Dataset

# Convertir cadenas "NAN" a verdaderos NaN
data.replace('NAN', np.nan, inplace=True)
#data = data.dropna()

data.shape

data=pd.DataFrame(data)

Productores_Res_En_Vereda=data[[Variable,'PRODUCTORES RESIDENTES']]
Productores_Res_En_Vereda.loc[:, 'PRODUCTORES RESIDENTES'] = Productores_Res_En_Vereda['PRODUCTORES RESIDENTES'].astype(float)
Productores_Res_En_Vereda=Productores_Res_En_Vereda.groupby(Variable).mean()
Productores_Res_En_Vereda=Productores_Res_En_Vereda.rename(columns={'PRODUCTORES RESIDENTES': "Productores_Res_En_Vereda"})

Unidad_Prod_No_Agro=data[[Variable,'UPNA']]
Unidad_Prod_No_Agro.loc[:, 'UPNA'] = Unidad_Prod_No_Agro['UPNA'].astype(float)
Unidad_Prod_No_Agro=Unidad_Prod_No_Agro.groupby(Variable).mean()
Unidad_Prod_No_Agro=Unidad_Prod_No_Agro.rename(columns={'UPNA': "Unidad_Prod_No_Agro"})

Unidad_Prod_Agro=data[[Variable,'UPA']]
Unidad_Prod_Agro.loc[:, 'UPA'] = Unidad_Prod_Agro['UPA'].astype(float)
Unidad_Prod_Agro=Unidad_Prod_Agro.groupby(Variable).mean()
Unidad_Prod_Agro=Unidad_Prod_Agro.rename(columns={'UPA': "Unidad_Prod_Agro"})

Unidad_Productiva=data[[Variable,'UP']]
Unidad_Productiva.loc[:, 'UP'] = Unidad_Productiva['UP'].astype(float)
Unidad_Productiva=Unidad_Productiva.groupby(Variable).mean()
Unidad_Productiva=Unidad_Productiva.rename(columns={'UP': "Unidad_Productiva"})

Personas_Vereda=data[[Variable,'PERSONAS']]
Personas_Vereda.loc[:, 'PERSONAS'] = Personas_Vereda['PERSONAS'].astype(float)
Personas_Vereda=Personas_Vereda.groupby(Variable).mean()
Personas_Vereda=Personas_Vereda.rename(columns={'PERSONAS': "Personas_Vereda"})

Hogares_Vereda=data[[Variable,'HOGARES']]
Hogares_Vereda.loc[:, 'HOGARES'] = Hogares_Vereda['HOGARES'].astype(float)
Hogares_Vereda=Hogares_Vereda.groupby(Variable).mean()
Hogares_Vereda=Hogares_Vereda.rename(columns={'HOGARES': "Hogares_Vereda"})

Viviendas_Veredas=data[[Variable,'VIVIENDAS']]
Viviendas_Veredas.loc[:, 'VIVIENDAS'] = Viviendas_Veredas['VIVIENDAS'].astype(float)
Viviendas_Veredas=Viviendas_Veredas.groupby(Variable).mean()
Viviendas_Veredas=Viviendas_Veredas.rename(columns={'VIVIENDAS': "Viviendas_Veredas"})

Area_HA_Vereda=data[[Variable,'AREA_HA']]
Area_HA_Vereda.loc[:, 'AREA_HA'] = Area_HA_Vereda['AREA_HA'].astype(float)
Area_HA_Vereda=Area_HA_Vereda.groupby(Variable).mean()
Area_HA_Vereda=Area_HA_Vereda.rename(columns={'AREA_HA': "Area_HA_Vereda"})

Medidas_Contra_HP=data[[Variable,'¿ha tomado alguna medida para evitar contraer esta infección?']]
Medidas_Contra_HP.replace('NOHAESCUCHADODELHELICOBACTERPYLORI', 'NO', inplace=True)
Medidas_Contra_HP.loc[:, '¿ha tomado alguna medida para evitar contraer esta infección?'] = Medidas_Contra_HP['¿ha tomado alguna medida para evitar contraer esta infección?'].astype('category').cat.codes
Medidas_Contra_HP=Medidas_Contra_HP.groupby(Variable).mean()
Medidas_Contra_HP=Medidas_Contra_HP.rename(columns={'¿ha tomado alguna medida para evitar contraer esta infección?': "Medidas_Contra_HP"})

Uso_Recipiente_Para_Agua=data[[Variable,'\xa0¿Utiliza un recipiente con tapa para almacenar el agua de consumo?']]
Uso_Recipiente_Para_Agua.loc[:, '\xa0¿Utiliza un recipiente con tapa para almacenar el agua de consumo?'] = Uso_Recipiente_Para_Agua['\xa0¿Utiliza un recipiente con tapa para almacenar el agua de consumo?'].astype('category').cat.codes
Uso_Recipiente_Para_Agua=Uso_Recipiente_Para_Agua.groupby(Variable).mean()
Uso_Recipiente_Para_Agua=Uso_Recipiente_Para_Agua.rename(columns={'\xa0¿Utiliza un recipiente con tapa para almacenar el agua de consumo?': "Uso_Recipiente_Para_Agua"})

Promedio_Dias_Con_Agua=data[[Variable,'¿De los 7 días de la semana cuantos días llega el agua?']]
Promedio_Dias_Con_Agua.loc[:, '¿De los 7 días de la semana cuantos días llega el agua?'] = Promedio_Dias_Con_Agua['¿De los 7 días de la semana cuantos días llega el agua?'].astype('category').cat.codes
Promedio_Dias_Con_Agua=Promedio_Dias_Con_Agua.groupby(Variable).mean()
Promedio_Dias_Con_Agua=Promedio_Dias_Con_Agua.rename(columns={'¿De los 7 días de la semana cuantos días llega el agua?': "Promedio_Dias_Con_Agua"})

Altitud=data[[Variable,'Altitud (msnm):']]
Altitud.loc[:, 'Altitud (msnm):'] = Altitud['Altitud (msnm):'].astype(float)
Altitud=Altitud.groupby(Variable).mean()
Altitud=Altitud.rename(columns={'Altitud (msnm):': "altitud"})

Tipo_Vivienda=data[[Variable, '¿Qué tipo de vivienda es?']]
Tipo_Vivienda.loc[:, '¿Qué tipo de vivienda es?'] = Tipo_Vivienda[ '¿Qué tipo de vivienda es?'].astype('category').cat.codes
Tipo_Vivienda=Tipo_Vivienda.groupby(Variable).agg(lambda x: pd.Series.mode(x)[0])
Tipo_Vivienda=Tipo_Vivienda.rename(columns={'¿Qué tipo de vivienda es?': "Tipo_Vivienda"})

MaterialTecho=data[[Variable,'¿Cuál es el material predominante del techo de la vivienda?']]
MaterialTecho.loc[:, '¿Cuál es el material predominante del techo de la vivienda?'] = MaterialTecho['¿Cuál es el material predominante del techo de la vivienda?'].astype('category').cat.codes
MaterialTecho=MaterialTecho.groupby(Variable).agg(lambda x: pd.Series.mode(x)[0])
MaterialTecho=MaterialTecho.rename(columns={'¿Cuál es el material predominante del techo de la vivienda?': "MaterialTecho"})

MaterialPiso=data[[Variable,'¿Cuál es el material predominante de los pisos de la vivienda?']]
MaterialPiso.loc[:, '¿Cuál es el material predominante de los pisos de la vivienda?'] = MaterialPiso['¿Cuál es el material predominante de los pisos de la vivienda?'].astype('category').cat.codes
MaterialPiso=MaterialPiso.groupby(Variable).agg(lambda x: pd.Series.mode(x)[0])
MaterialPiso=MaterialPiso.rename(columns={'¿Cuál es el material predominante de los pisos de la vivienda?': "MaterialPiso"})

MaterialPared=data[[Variable,'¿Cuál es el material predominante de las paredes exteriores?']]
MaterialPared.loc[:, '¿Cuál es el material predominante de las paredes exteriores?'] = MaterialPared['¿Cuál es el material predominante de las paredes exteriores?'].astype('category').cat.codes
MaterialPared=MaterialPared.groupby(Variable).agg(lambda x: pd.Series.mode(x)[0])
MaterialPared=MaterialPared.rename(columns={'¿Cuál es el material predominante de las paredes exteriores?': "MaterialPared"})

Alcantarillado=data[[Variable,'¿La unidad de vivienda cuenta con servicio de alcantarillado?']]
Alcantarillado.loc[:, '¿La unidad de vivienda cuenta con servicio de alcantarillado?'] = Alcantarillado['¿La unidad de vivienda cuenta con servicio de alcantarillado?'].astype('category').cat.codes
Alcantarillado=Alcantarillado.groupby(Variable).mean()
Alcantarillado=Alcantarillado.rename(columns={'¿La unidad de vivienda cuenta con servicio de alcantarillado?': "Alcantarillado"})

Basura=data[[Variable,'¿Que método emplea\xa0principalmente\xa0para eliminar las basuras en este hogar?']]
Basura.loc[:, '¿Que método emplea\xa0principalmente\xa0para eliminar las basuras en este hogar?'] = Basura['¿Que método emplea\xa0principalmente\xa0para eliminar las basuras en este hogar?'].astype('category').cat.codes
Basura=Basura.groupby(Variable).agg(lambda x: pd.Series.mode(x)[0])
Basura=Basura.rename(columns={'¿Que método emplea\xa0principalmente\xa0para eliminar las basuras en este hogar?': "Basura"})

Familias_Usando_Cocina=data[[Variable,'¿Cuantos grupos de personas que cocinen por separado hay en el hogar?']]
Familias_Usando_Cocina.loc[:, '¿Cuantos grupos de personas que cocinen por separado hay en el hogar?'] = Familias_Usando_Cocina['¿Cuantos grupos de personas que cocinen por separado hay en el hogar?'].astype(float)
Familias_Usando_Cocina=Familias_Usando_Cocina.groupby(Variable).mean()
Familias_Usando_Cocina=Familias_Usando_Cocina.rename(columns={'¿Cuantos grupos de personas que cocinen por separado hay en el hogar?': "Familias_Usando_Cocina"})

habitantes=data[[Variable,'¿Cuántas personas habitan la unidad de vivienda?']]
habitantes = habitantes.astype({"¿Cuántas personas habitan la unidad de vivienda?": float})
habitantes=habitantes.groupby(Variable).mean()
habitantes=habitantes.rename(columns={"¿Cuántas personas habitan la unidad de vivienda?": "Habitantes/Vivienda"})

Etnia=data[[Variable,'¿Grupo étnico al cual pertenece?']]
Etnia.loc[:, '¿Grupo étnico al cual pertenece?'] = Etnia['¿Grupo étnico al cual pertenece?'].astype('category').cat.codes
Etnia=Etnia.groupby(Variable).agg(lambda x: pd.Series.mode(x)[0])
Etnia=Etnia.rename(columns={'¿Grupo étnico al cual pertenece?': "Etnia"})

Lav_Post_San=data[[Variable,'¿En el hogar se tiene el habito de lavar sus manos después de ir al servicio sanitario?']]
Lav_Post_San.loc[:, '¿En el hogar se tiene el habito de lavar sus manos después de ir al servicio sanitario?'] = Lav_Post_San['¿En el hogar se tiene el habito de lavar sus manos después de ir al servicio sanitario?'].astype('category').cat.codes
Lav_Post_San=Lav_Post_San.groupby(Variable).mean()
Lav_Post_San=Lav_Post_San.rename(columns={'¿En el hogar se tiene el habito de lavar sus manos después de ir al servicio sanitario?': "LavadoPostSanitario"})

Sanitarios=data[[Variable,'¿Cuantos servicios sanitarios o inodoros dispone este hogar?']]
Sanitarios.loc[:, '¿Cuantos servicios sanitarios o inodoros dispone este hogar?'] = Sanitarios['¿Cuantos servicios sanitarios o inodoros dispone este hogar?'].astype(float)
Sanitarios=Sanitarios.groupby(Variable).mean()
Sanitarios=Sanitarios.rename(columns={"¿Cuantos servicios sanitarios o inodoros dispone este hogar?": "Cantidad_Sanitarios"})

Tipo_Sanitarios=data[[Variable,'¿Qué servicio sanitario se utiliza en el hogar?']]
Tipo_Sanitarios.loc[:,'¿Qué servicio sanitario se utiliza en el hogar?'] = Tipo_Sanitarios['¿Qué servicio sanitario se utiliza en el hogar?'].astype('category').cat.codes
Tipo_Sanitarios=Tipo_Sanitarios.groupby(Variable).agg(lambda x: pd.Series.mode(x)[0])
Tipo_Sanitarios=Tipo_Sanitarios.rename(columns={ '¿Qué servicio sanitario se utiliza en el hogar?': "Tipo_Sanitarios"})

Ducha=data[[Variable,'¿Tiene ducha o regadera conectada al acueducto?']]
Ducha.loc[:, '¿Tiene ducha o regadera conectada al acueducto?'] = Ducha['¿Tiene ducha o regadera conectada al acueducto?'].astype('category').cat.codes
Ducha=Ducha.groupby(Variable).mean()
Ducha=Ducha.rename(columns={'¿Tiene ducha o regadera conectada al acueducto?': "Ducha"})

Lav_Ant_Comer=data[[Variable,'¿Los integrantes del hogar tienen el hábito de lavar sus manos antes de comer?']]
Lav_Ant_Comer.loc[:, '¿Los integrantes del hogar tienen el hábito de lavar sus manos antes de comer?'] = Lav_Ant_Comer['¿Los integrantes del hogar tienen el hábito de lavar sus manos antes de comer?'].astype('category').cat.codes
Lav_Ant_Comer=Lav_Ant_Comer.groupby(Variable).mean()
Lav_Ant_Comer=Lav_Ant_Comer.rename(columns={'¿Los integrantes del hogar tienen el hábito de lavar sus manos antes de comer?': "LavadoAntesComer"})

Sedimentos=data[[Variable,'Sedimentos:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?']]
Sedimentos.loc[:, 'Sedimentos:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?'] = Sedimentos['Sedimentos:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?'].astype('category').cat.codes
Sedimentos=Sedimentos.groupby(Variable).mean()
Sedimentos=Sedimentos.rename(columns={'Sedimentos:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?': "Sedimentos"})

Mal_Sabor=data[[Variable,'Mal sabor:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?']]
Mal_Sabor.loc[:, 'Mal sabor:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?'] = Mal_Sabor['Mal sabor:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?'].astype('category').cat.codes
Mal_Sabor=Mal_Sabor.groupby(Variable).mean()
Mal_Sabor=Mal_Sabor.rename(columns={'Mal sabor:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?': "Mal_Sabor"})

Mal_Color=data[[Variable,'Mal color:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?']]
Mal_Color.loc[:, 'Mal color:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?'] = Mal_Color['Mal color:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?'].astype('category').cat.codes
Mal_Color=Mal_Color.groupby(Variable).mean()
Mal_Color=Mal_Color.rename(columns={'Mal color:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?': "Mal_Color"})

Mal_Olor=data[[Variable,'Mal olor:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?']]
Mal_Olor.loc[:, 'Mal olor:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?'] = Mal_Olor['Mal olor:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?'].astype('category').cat.codes
Mal_Olor=Mal_Olor.groupby(Variable).mean()
Mal_Olor=Mal_Olor.rename(columns={'Mal olor:¿El agua utilizada por el hogar para preparar los alimentos y para consumo presenta las siguientes características?': "Mal_Olor"})

Purificacion=data[[Variable,'¿Qué método emplea principalmente para purificar el  agua para consumo?']]
Purificacion.loc[:, '¿Qué método emplea principalmente para purificar el  agua para consumo?'] = Purificacion['¿Qué método emplea principalmente para purificar el  agua para consumo?'].astype('category').cat.codes
Purificacion=Purificacion.groupby(Variable).agg(lambda x: pd.Series.mode(x)[0])
Purificacion=Purificacion.rename(columns={'¿Qué método emplea principalmente para purificar el  agua para consumo?': "Purificacion"})

Cepillado=data[[Variable,'¿Cuántas veces se cepillan los dientes al día?']]
Cepillado= Cepillado.astype({"¿Cuántas veces se cepillan los dientes al día?": float})
Cepillado=Cepillado.groupby(Variable).mean()
Cepillado=Cepillado.rename(columns={'¿Cuántas veces se cepillan los dientes al día?': "Cepillado"})

AnimalDomestico=data[[Variable,'¿Tienen en la unidad de vivienda animales domésticos?']]
AnimalDomestico.loc[:, '¿Tienen en la unidad de vivienda animales domésticos?'] = AnimalDomestico['¿Tienen en la unidad de vivienda animales domésticos?'].astype('category').cat.codes
AnimalDomestico=AnimalDomestico.groupby(Variable).mean()
AnimalDomestico=AnimalDomestico.rename(columns={'¿Tienen en la unidad de vivienda animales domésticos?': "AnimalDomestico"})

Cuartos_Vivienda=data[[Variable, '¿Cuántos cuartos o piezas Incluyendo sala y comedor, dispone esta unidad de vivienda?']]
Cuartos_Vivienda = Cuartos_Vivienda.astype({"¿Cuántos cuartos o piezas Incluyendo sala y comedor, dispone esta unidad de vivienda?": float})
Cuartos_Vivienda=Cuartos_Vivienda.groupby(Variable).mean()
Cuartos_Vivienda=Cuartos_Vivienda.rename(columns={'¿Cuántos cuartos o piezas Incluyendo sala y comedor, dispone esta unidad de vivienda?': "Cuartos/Vivienda"})

Dormitorios=data[[Variable,'¿En cuántos de esos cuartos duermen las personas de este hogar?']]
Dormitorios.loc[:,'¿En cuántos de esos cuartos duermen las personas de este hogar?'] = Dormitorios['¿En cuántos de esos cuartos duermen las personas de este hogar?'].astype(float)
Dormitorios=Dormitorios.groupby(Variable).mean()
Dormitorios=Dormitorios.rename(columns={'¿En cuántos de esos cuartos duermen las personas de este hogar?': "Dormitorios/Vivienda"})

Ubicacion_Sanitario=data[[Variable, '¿En dónde está ubicado el servicio sanitario que utiliza el hogar?']]
Ubicacion_Sanitario.loc[:, '¿En dónde está ubicado el servicio sanitario que utiliza el hogar?'] = Ubicacion_Sanitario[ '¿En dónde está ubicado el servicio sanitario que utiliza el hogar?'].astype('category').cat.codes
Ubicacion_Sanitario=Ubicacion_Sanitario.groupby(Variable).agg(lambda x: pd.Series.mode(x)[0])
Ubicacion_Sanitario=Ubicacion_Sanitario.rename(columns={'¿En dónde está ubicado el servicio sanitario que utiliza el hogar?': "Ubicacion_Sanitario"})

Fuente_Agua=data[[Variable, '¿De dónde obtiene principalmente este hogar el agua para consumo humano?']]
Fuente_Agua.loc[:, '¿De dónde obtiene principalmente este hogar el agua para consumo humano?'] = Fuente_Agua[ '¿De dónde obtiene principalmente este hogar el agua para consumo humano?'].astype('category').cat.codes
Fuente_Agua=Fuente_Agua.groupby(Variable).agg(lambda x: pd.Series.mode(x)[0])
Fuente_Agua=Fuente_Agua.rename(columns={'¿De dónde obtiene principalmente este hogar el agua para consumo humano?': "Fuente_Agua"})

Cocina_Exclusiva=data[[Variable,'¿El hogar tiene cocina (espacio exclusivo para preparar alimentos)?']]
Cocina_Exclusiva.loc[:,'¿El hogar tiene cocina (espacio exclusivo para preparar alimentos)?'] = Cocina_Exclusiva['¿El hogar tiene cocina (espacio exclusivo para preparar alimentos)?'].astype('category').cat.codes
Cocina_Exclusiva=Cocina_Exclusiva.groupby(Variable).mean()
Cocina_Exclusiva=Cocina_Exclusiva.rename(columns={'¿El hogar tiene cocina (espacio exclusivo para preparar alimentos)?': "Cocina_Exclusiva"})

Dias_Ducha=data[[Variable,'¿Cuantos días a la semana se bañan los miembros de la familia?']]
Dias_Ducha.loc[:,'¿Cuantos días a la semana se bañan los miembros de la familia?'] = Dias_Ducha['¿Cuantos días a la semana se bañan los miembros de la familia?'].astype('category').cat.codes
Dias_Ducha=Dias_Ducha.groupby(Variable).mean()
Dias_Ducha=Dias_Ducha.rename(columns={'¿Cuantos días a la semana se bañan los miembros de la familia?': "Dias_Ducha"})

Fuente_Combustible=data[[Variable,'¿Con qué energía o combustible cocinan principalmente en este hogar?']]
Fuente_Combustible.loc[:,'¿Con qué energía o combustible cocinan principalmente en este hogar?'] = Fuente_Combustible['¿Con qué energía o combustible cocinan principalmente en este hogar?'].astype('category').cat.codes
Fuente_Combustible=Fuente_Combustible.groupby(Variable).agg(lambda x: pd.Series.mode(x)[0])
Fuente_Combustible=Fuente_Combustible.rename(columns={'¿Con qué energía o combustible cocinan principalmente en este hogar?': "Fuente_Combustible"})

Comparten_Cepillo=data[[Variable,'¿Comparten los miembros del hogar el cepillo de dientes?']]
Comparten_Cepillo.loc[:,'¿Comparten los miembros del hogar el cepillo de dientes?'] = Comparten_Cepillo['¿Comparten los miembros del hogar el cepillo de dientes?'].astype('category').cat.codes
Comparten_Cepillo=Comparten_Cepillo.groupby(Variable).mean()
Comparten_Cepillo=Comparten_Cepillo.rename(columns={'¿Comparten los miembros del hogar el cepillo de dientes?': "Comparten_Cepillo"})

Comparten_Cubiero=data[[Variable,'¿Comparten los miembros del hogar cubiertos y/o platos sin previo lavado de los mismos?']]
Comparten_Cubiero.loc[:,'¿Comparten los miembros del hogar cubiertos y/o platos sin previo lavado de los mismos?'] = Comparten_Cubiero['¿Comparten los miembros del hogar cubiertos y/o platos sin previo lavado de los mismos?'].astype('category').cat.codes
Comparten_Cubiero=Comparten_Cubiero.groupby(Variable).mean()
Comparten_Cubiero=Comparten_Cubiero.rename(columns={'¿Comparten los miembros del hogar cubiertos y/o platos sin previo lavado de los mismos?': "Comparten_Cubiero"})

Contaco_Animales_Granja=data[[Variable,'¿Usted o algún integrante del hogar cría animales de granja (gallinas, patos, conejos, cerdos etc.)?']]
Contaco_Animales_Granja.loc[:,'¿Usted o algún integrante del hogar cría animales de granja (gallinas, patos, conejos, cerdos etc.)?'] = Contaco_Animales_Granja['¿Usted o algún integrante del hogar cría animales de granja (gallinas, patos, conejos, cerdos etc.)?'].astype('category').cat.codes
Contaco_Animales_Granja=Contaco_Animales_Granja.groupby(Variable).mean()
Contaco_Animales_Granja=Contaco_Animales_Granja.rename(columns={'¿Usted o algún integrante del hogar cría animales de granja (gallinas, patos, conejos, cerdos etc.)?': "Contaco_Animales_Granja"})

Convivencia_infeccion=data[[Variable,'g_factores/convive_personas_infectadas']]
Convivencia_infeccion.loc[:,'g_factores/convive_personas_infectadas'] = Convivencia_infeccion['g_factores/convive_personas_infectadas'].astype('category').cat.codes
Convivencia_infeccion=Convivencia_infeccion.groupby(Variable).mean()
Convivencia_infeccion=Convivencia_infeccion.rename(columns={'g_factores/convive_personas_infectadas': "Convivencia_infeccion"})

Antesedentes_Gastritis=data[[Variable,'Gastritis:¿Alguno de sus familiares ha sido diagnosticado con alguna de estas enfermedades?']]
Antesedentes_Gastritis.loc[:,'Gastritis:¿Alguno de sus familiares ha sido diagnosticado con alguna de estas enfermedades?'] = Antesedentes_Gastritis['Gastritis:¿Alguno de sus familiares ha sido diagnosticado con alguna de estas enfermedades?'].astype('category').cat.codes
Antesedentes_Gastritis=Antesedentes_Gastritis.groupby(Variable).mean()
Antesedentes_Gastritis=Antesedentes_Gastritis.rename(columns={'Gastritis:¿Alguno de sus familiares ha sido diagnosticado con alguna de estas enfermedades?': "Antesedentes_Gastritis"})

Antesedentes_Ulceras=data[[Variable,'Úlceras:¿Alguno de sus familiares ha sido diagnosticado con alguna de estas enfermedades?']]
Antesedentes_Ulceras.loc[:,'Úlceras:¿Alguno de sus familiares ha sido diagnosticado con alguna de estas enfermedades?'] = Antesedentes_Ulceras['Úlceras:¿Alguno de sus familiares ha sido diagnosticado con alguna de estas enfermedades?'].astype('category').cat.codes
Antesedentes_Ulceras=Antesedentes_Ulceras.groupby(Variable).mean()
Antesedentes_Ulceras=Antesedentes_Ulceras.rename(columns={'Úlceras:¿Alguno de sus familiares ha sido diagnosticado con alguna de estas enfermedades?': "Antesedentes_Ulceras"})

Antesedentes_CG=data[[Variable,'Cáncer gástrico:¿Alguno de sus familiares ha sido diagnosticado con alguna de estas enfermedades?']]
Antesedentes_CG.loc[:,'Cáncer gástrico:¿Alguno de sus familiares ha sido diagnosticado con alguna de estas enfermedades?'] = Antesedentes_CG['Cáncer gástrico:¿Alguno de sus familiares ha sido diagnosticado con alguna de estas enfermedades?'].astype('category').cat.codes
Antesedentes_CG=Antesedentes_CG.groupby(Variable).mean()
Antesedentes_CG=Antesedentes_CG.rename(columns={'Cáncer gástrico:¿Alguno de sus familiares ha sido diagnosticado con alguna de estas enfermedades?': "Antesedentes_CG"})

Ratas=data[[Variable,'g_factores/plagas/Ratas']]
Ratas.loc[:,'g_factores/plagas/Ratas'] = Ratas['g_factores/plagas/Ratas'].astype('category').cat.codes
Ratas=Ratas.groupby(Variable).mean()
Ratas=Ratas.rename(columns={'g_factores/plagas/Ratas': "Ratas"})

Cucarachas=data[[Variable,'g_factores/plagas/Cucarachas']]
Cucarachas.loc[:,'g_factores/plagas/Cucarachas'] = Cucarachas['g_factores/plagas/Cucarachas'].astype('category').cat.codes
Cucarachas=Cucarachas.groupby(Variable).mean()
Cucarachas=Cucarachas.rename(columns={'g_factores/plagas/Cucarachas': "Cucarachas"})

Murcielagos=data[[Variable,'g_factores/plagas/Murcielagos']]
Murcielagos.loc[:,'g_factores/plagas/Murcielagos'] = Murcielagos['g_factores/plagas/Murcielagos'].astype('category').cat.codes
Murcielagos=Murcielagos.groupby(Variable).mean()
Murcielagos=Murcielagos.rename(columns={'g_factores/plagas/Murcielagos': "Murcielagos"})

Zancudos=data[[Variable,'g_factores/plagas/Zancudos']]
Zancudos.loc[:,'g_factores/plagas/Zancudos'] = Zancudos['g_factores/plagas/Zancudos'].astype('category').cat.codes
Zancudos=Zancudos.groupby(Variable).mean()
Zancudos=Zancudos.rename(columns={'g_factores/plagas/Zancudos': "Zancudos"})

# Trabajo de altura sobre vereda

Latitud=data[[Variable,'Latitud']]
Latitud=Latitud.groupby(Variable).agg(lambda x: pd.Series.mode(x)[0])

Longitud=data[[Variable,'Longitud']]
Longitud=Longitud.groupby(Variable).agg(lambda x: pd.Series.mode(x)[0])

import requests

# Define tus API keys
OWM_API_KEY = '6ecddbe5a9490401ed9b8e005db03ba0'
temperaturas = []
humedades = []
altitudes = []

dfMap=Latitud.join(Longitud)

# Listas para almacenar los resultados
# Itera sobre cada fila del DataFrame
for index, row in dfMap.iterrows():
    lat, lon = row['Latitud'], row['Longitud']
    # Consulta OpenWeatherMap para la temperatura
    weather_url = f"http://api.openweathermap.org/data/2.5/weather?lat={lat}&lon={lon}&appid={OWM_API_KEY}&units=metric"
    weather_response = requests.get(weather_url)
    weather_data = weather_response.json()
    temperatura = weather_data['main']['temp']
    humedad= weather_data['main']['humidity']
    temperaturas.append(temperatura)
    humedades.append(humedad)


# Añadir los datos al DataFrame original
dfMap['Humedades']= humedades
dfMap['Temperatura'] = temperaturas

# Itera sobre cada fila del DataFrame
for index, row in dfMap.iterrows():
   lat, lon = row['Latitud'], row['Longitud']
   elevation_url = f"https://api.open-elevation.com/api/v1/lookup?locations={lat},{lon}"
   elevation_response = requests.get(elevation_url)
   elevation_data = elevation_response.json()
   altitud = elevation_data['results'][0]['elevation']
   altitudes.append(altitud)

# Añadir los datos al DataFrame original
dfMap['Altitud'] = altitudes

import numpy as np
dfMap=dfMap.join(Altitud)
dfMap['Altitud_Vereda'] = np.where(dfMap['altitud'] == 0, dfMap['Altitud'], dfMap['altitud'])
dfMap=dfMap.drop(['altitud','Altitud'], axis=1 )

## Creacion del Dataset

k=['3','5','7','10']
for i in k:

  prevalencia=positivos.join(total)
  prevalencia.fillna(0, inplace=True)
  prevalencia['Prevalencia']=prevalencia['Positivos']/prevalencia['Total']

  kbins = KBinsDiscretizer(n_bins= int(i), encode='ordinal', strategy='uniform')
  label_ids=prevalencia['Prevalencia'].to_numpy()
  reshaped=label_ids.reshape(len(label_ids), 1)
  data_trans = kbins.fit_transform(reshaped)
  #result = np.stack((prevalencia.index,data_trans.ravel()), axis=-1)

  prevalencia=pd.DataFrame(data_trans,index=prevalencia.index.ravel(),columns=["Prevalencia"])

  dataset=MaterialTecho.join(habitantes)
  dataset=dataset.join(dfMap)
  dataset=dataset.join(Uso_Recipiente_Para_Agua)
  dataset=dataset.join(Medidas_Contra_HP)
  dataset=dataset.join(Area_HA_Vereda)
  dataset=dataset.join(Viviendas_Veredas)
  dataset=dataset.join(Hogares_Vereda)
  dataset=dataset.join(Personas_Vereda)
  dataset=dataset.join(Unidad_Productiva)
  dataset=dataset.join(Unidad_Prod_Agro)
  dataset=dataset.join(Unidad_Prod_No_Agro)
  dataset=dataset.join(Productores_Res_En_Vereda)
  dataset=dataset.join(Promedio_Dias_Con_Agua)
  dataset=dataset.join(Zancudos)
  dataset=dataset.join(Ratas)
  dataset=dataset.join(Cucarachas)
  dataset=dataset.join(Murcielagos)
  dataset=dataset.join(Antesedentes_CG)
  dataset=dataset.join(Antesedentes_Ulceras)
  dataset=dataset.join(Antesedentes_Gastritis)
  dataset=dataset.join(Convivencia_infeccion)
  dataset=dataset.join(Contaco_Animales_Granja)
  dataset=dataset.join(Comparten_Cubiero)
  dataset=dataset.join(Comparten_Cepillo)
  dataset=dataset.join(Fuente_Combustible)
  dataset=dataset.join(Dias_Ducha)
  dataset=dataset.join(Cocina_Exclusiva)
  dataset=dataset.join(Fuente_Agua)
  dataset=dataset.join(Ubicacion_Sanitario)
  dataset=dataset.join(Dormitorios)
  dataset=dataset.join(Cuartos_Vivienda)
  dataset=dataset.join(Tipo_Vivienda)
  dataset=dataset.join(Familias_Usando_Cocina)
  dataset=dataset.join(Alcantarillado)
  dataset=dataset.join(AnimalDomestico)
  dataset=dataset.join(Cepillado)
  dataset=dataset.join(Purificacion)
  dataset=dataset.join(Mal_Olor)
  dataset=dataset.join(Mal_Color)
  dataset=dataset.join(Mal_Sabor)
  dataset=dataset.join(Sedimentos)
  dataset=dataset.join(Lav_Ant_Comer)
  dataset=dataset.join(Ducha)
  dataset=dataset.join(Basura)
  dataset=dataset.join(MaterialPiso)
  dataset=dataset.join(MaterialPared)
  dataset=dataset.join(Tipo_Sanitarios)
  dataset=dataset.join(Sanitarios)
  dataset=dataset.join(Lav_Post_San)
  dataset=dataset.join(Etnia)
  dataset=dataset.join(prevalencia['Prevalencia'])

  for column in dataset.columns:
    try:
        dataset[column] = dataset[column].astype(float)
    except ValueError:
        print(f"La columna {column} no se puede convertir a float.")

  dataset = dataset.round(5)
  dataset=dataset.dropna()

  dataset.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Prevalencias/Prevalencia'+i+'.csv')

dataset.shape
