==> 2_Tratamiento_csv.ipynb <==
!pip install -U scikit-learn==1.2.2

import numpy as np
import pandas as pd
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.feature_selection import chi2
from sklearn.feature_selection import SelectKBest, mutual_info_classif

from google.colab import drive
drive.mount('/content/drive')

# Valores de Prevalencia en rango 0 a 2 (3 valores)

prevalencia = pd.read_csv("/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Prevalencias/Prevalencia3.csv", sep=",",header=0, index_col=0)

prevalencia=prevalencia.reset_index(drop=True)
dataset=pd.DataFrame(prevalencia)

dfSet=dataset.drop(['Longitud','Latitud'], axis=1)
dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel8Pre3.csv')
dfSet=dfSet.drop(['Unidad_Prod_Agro','Unidad_Prod_No_Agro','Unidad_Productiva','Productores_Res_En_Vereda'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel7Pre3.csv')
dfSet=dfSet.drop(['Humedades',	'Temperatura',	'Altitud_Vereda', ], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel6Pre3.csv')
dfSet=dfSet.drop(['Area_HA_Vereda', 'Alcantarillado', 'Viviendas_Veredas', 'Hogares_Vereda', 'Personas_Vereda', 'Promedio_Dias_Con_Agua'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel5Pre3.csv')
dfSet=dfSet.drop(['Medidas_Contra_HP','LavadoPostSanitario','Antesedentes_CG', 'LavadoAntesComer', 'Cepillado', 'Etnia','Antesedentes_Ulceras',
                  'Antesedentes_Gastritis','Convivencia_infeccion', 'Dias_Ducha', 'Uso_Recipiente_Para_Agua',  'Comparten_Cubiero','Comparten_Cepillo',], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel4Pre3.csv')
dfSet=dfSet.drop(['AnimalDomestico','Contaco_Animales_Granja','Zancudos',
       'Ratas', 'Cucarachas', 'Murcielagos',], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel3Pre3.csv')
dfSet=dfSet.drop(['Tipo_Vivienda', 'Cantidad_Sanitarios','Fuente_Combustible','Ducha', 'Basura','MaterialPiso', 'MaterialPared','Tipo_Sanitarios','MaterialTecho',
                  'Cocina_Exclusiva','Fuente_Agua','Ubicacion_Sanitario', 'Dormitorios/Vivienda', 'Cuartos/Vivienda','Familias_Usando_Cocina'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel2Pre3.csv')
dfSet=dfSet.drop(['Habitantes/Vivienda'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel1Pre3.csv')
dfSet=dfSet.drop(['Purificacion'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel0Pre3.csv')

# Valores de Prevalencia en rango 0 a 4 (5 valores)

prevalencia = pd.read_csv("/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Prevalencias/Prevalencia5.csv", sep=",",header=0, index_col=0)

prevalencia=prevalencia.reset_index(drop=True)
dataset=pd.DataFrame(prevalencia)

dfSet=dataset.drop(['Longitud','Latitud'], axis=1)
dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel8Pre5.csv')
dfSet=dfSet.drop(['Unidad_Prod_Agro','Unidad_Prod_No_Agro','Unidad_Productiva','Productores_Res_En_Vereda'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel7Pre5.csv')
dfSet=dfSet.drop(['Humedades',	'Temperatura',	'Altitud_Vereda', ], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel6Pre5.csv')
dfSet=dfSet.drop(['Area_HA_Vereda', 'Alcantarillado', 'Viviendas_Veredas', 'Hogares_Vereda', 'Personas_Vereda', 'Promedio_Dias_Con_Agua'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel5Pre5.csv')
dfSet=dfSet.drop(['Medidas_Contra_HP','LavadoPostSanitario','Antesedentes_CG', 'LavadoAntesComer', 'Cepillado', 'Etnia','Antesedentes_Ulceras', 'Antesedentes_Gastritis','Convivencia_infeccion', 'Dias_Ducha', 'Uso_Recipiente_Para_Agua',  'Comparten_Cubiero','Comparten_Cepillo',], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel4Pre5.csv')
dfSet=dfSet.drop(['AnimalDomestico','Contaco_Animales_Granja','Zancudos',
       'Ratas', 'Cucarachas', 'Murcielagos',], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel3Pre5.csv')
dfSet=dfSet.drop(['Tipo_Vivienda', 'Cantidad_Sanitarios','Fuente_Combustible','Ducha', 'Basura','MaterialPiso', 'MaterialPared','Tipo_Sanitarios','MaterialTecho','Cocina_Exclusiva','Fuente_Agua','Ubicacion_Sanitario', 'Dormitorios/Vivienda', 'Cuartos/Vivienda','Familias_Usando_Cocina'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel2Pre5.csv')
dfSet=dfSet.drop(['Habitantes/Vivienda'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel1Pre5.csv')
dfSet=dfSet.drop(['Purificacion'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel0Pre5.csv')

# Valores de Prevalencia en rango 0 a 6 (7 valores)

prevalencia = pd.read_csv("/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Prevalencias/Prevalencia7.csv", sep=",",header=0, index_col=0)

prevalencia=prevalencia.reset_index(drop=True)
dataset=pd.DataFrame(prevalencia)

dfSet=dataset.drop(['Longitud','Latitud'], axis=1)
dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel8Pre7.csv')
dfSet=dfSet.drop(['Unidad_Prod_Agro','Unidad_Prod_No_Agro','Unidad_Productiva','Productores_Res_En_Vereda'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel7Pre7.csv')
dfSet=dfSet.drop(['Humedades',	'Temperatura',	'Altitud_Vereda', ], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel6Pre7.csv')
dfSet=dfSet.drop(['Area_HA_Vereda', 'Alcantarillado', 'Viviendas_Veredas', 'Hogares_Vereda', 'Personas_Vereda', 'Promedio_Dias_Con_Agua'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel5Pre7.csv')
dfSet=dfSet.drop(['Medidas_Contra_HP','LavadoPostSanitario','Antesedentes_CG', 'LavadoAntesComer', 'Cepillado', 'Etnia','Antesedentes_Ulceras', 'Antesedentes_Gastritis','Convivencia_infeccion', 'Dias_Ducha', 'Uso_Recipiente_Para_Agua',  'Comparten_Cubiero','Comparten_Cepillo',], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel4Pre7.csv')
dfSet=dfSet.drop(['AnimalDomestico','Contaco_Animales_Granja','Zancudos',
       'Ratas', 'Cucarachas', 'Murcielagos',], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel3Pre7.csv')
dfSet=dfSet.drop(['Tipo_Vivienda', 'Cantidad_Sanitarios','Fuente_Combustible','Ducha', 'Basura','MaterialPiso', 'MaterialPared','Tipo_Sanitarios','MaterialTecho','Cocina_Exclusiva','Fuente_Agua','Ubicacion_Sanitario', 'Dormitorios/Vivienda', 'Cuartos/Vivienda','Familias_Usando_Cocina'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel2Pre7.csv')
dfSet=dfSet.drop(['Habitantes/Vivienda'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel1Pre7.csv')
dfSet=dfSet.drop(['Purificacion'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel0Pre7.csv')

# Valores de Prevalencia en rango 0 a 9 (10 valores)

prevalencia = pd.read_csv("/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Prevalencias/Prevalencia10.csv", sep=",",header=0, index_col=0)

prevalencia=prevalencia.reset_index(drop=True)
dataset=pd.DataFrame(prevalencia)

dfSet=dataset.drop(['Longitud','Latitud'], axis=1)
dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel8Pre10.csv')
dfSet=dfSet.drop(['Unidad_Prod_Agro','Unidad_Prod_No_Agro','Unidad_Productiva','Productores_Res_En_Vereda'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel7Pre10.csv')
dfSet=dfSet.drop(['Humedades',	'Temperatura',	'Altitud_Vereda', ], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel6Pre10.csv')
dfSet=dfSet.drop(['Area_HA_Vereda', 'Alcantarillado', 'Viviendas_Veredas', 'Hogares_Vereda', 'Personas_Vereda', 'Promedio_Dias_Con_Agua'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel5Pre10.csv')
dfSet=dfSet.drop(['Medidas_Contra_HP','LavadoPostSanitario','Antesedentes_CG', 'LavadoAntesComer', 'Cepillado', 'Etnia','Antesedentes_Ulceras', 'Antesedentes_Gastritis','Convivencia_infeccion', 'Dias_Ducha', 'Uso_Recipiente_Para_Agua',  'Comparten_Cubiero','Comparten_Cepillo',], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel4Pre10.csv')
dfSet=dfSet.drop(['AnimalDomestico','Contaco_Animales_Granja','Zancudos',
       'Ratas', 'Cucarachas', 'Murcielagos',], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel3Pre10.csv')
dfSet=dfSet.drop(['Tipo_Vivienda', 'Cantidad_Sanitarios','Fuente_Combustible','Ducha', 'Basura','MaterialPiso', 'MaterialPared','Tipo_Sanitarios','MaterialTecho','Cocina_Exclusiva','Fuente_Agua','Ubicacion_Sanitario', 'Dormitorios/Vivienda', 'Cuartos/Vivienda','Familias_Usando_Cocina'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel2Pre10.csv')
dfSet=dfSet.drop(['Habitantes/Vivienda'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel1Pre10.csv')
dfSet=dfSet.drop(['Purificacion'], axis=1)

dfSet.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev/Sel0Pre10.csv')
