==> 3_1_Redes_Neuronales_Ejecucion_Metodos.ipynb <==
!pip install -U scikit-learn==1.2.2

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import os

from google.colab import drive
from imblearn.over_sampling import SMOTE
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.neural_network import MLPClassifier
from sklearn.metrics import accuracy_score
from sklearn.pipeline import Pipeline
from imblearn.over_sampling import SMOTE
from sklearn.neural_network import MLPClassifier
from sklearn.metrics import accuracy_score
from sklearn.metrics import classification_report
from sklearn.preprocessing import label_binarize, StandardScaler
from sklearn.metrics import roc_curve, auc
from sklearn.decomposition import PCA
from sklearn.pipeline import Pipeline
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.svm import SVC
from sklearn.feature_selection import SelectKBest, f_classif, mutual_info_classif, chi2

drive.mount('/content/drive')

folder_path = '/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Var_Prev'
csv_files = [file for file in os.listdir(folder_path) if file.endswith('.csv')]

## Lista para almacenar dataframes
results = []
sm = SMOTE(random_state=42)

for file in csv_files:

    file_path = os.path.join(folder_path, file)
    df = pd.read_csv(file_path, sep=",", header=0, index_col=0)
    df = df.dropna()
    df = df.sample(n=500, replace=True, random_state=42)
    X = df.drop("Prevalencia", axis=1)
    y = df["Prevalencia"]

    # Determinar 'k', el número de variables
    k = X.shape[1]

    pipeline = Pipeline([
        ('scaler', StandardScaler()),    # Escalado de características
        ('reduce_dim', 'passthrough'),   # Placeholder para la reducción de dimensionalidad
        ('classification',MLPClassifier(random_state=42, verbose=False))
    ])

    # Definir los parámetros para GridSearch
    param_grid = [
        {
            'reduce_dim': [SelectKBest(score_func=f_classif),
                SelectKBest(score_func=mutual_info_classif),
                SelectKBest(score_func=chi2)],
            'reduce_dim__k': [k//2,k//4,k//8],
            'classification__learning_rate_init': [0.01,0.001,0.1],
            'classification__max_iter': [2000],
            'classification__momentum': [0.6,0.62,0.64],
            'classification__solver': ['adam'],
            'classification__hidden_layer_sizes':[ (4*k, 2*k),(2*k, 2*k),
             (4*k, 4*k),(4*k, 2*k,2*k),]
        },
        {
            'reduce_dim': [PCA(random_state=42)],
            'reduce_dim__n_components': [k//2,k//4,k//8],
            'classification__learning_rate_init': [0.01,0.001,0.1],
            'classification__max_iter': [2000],
            'classification__momentum': [0.6,0.62,0.64],
            'classification__solver': ['adam'],
            'classification__hidden_layer_sizes':[ (4*k, 2*k),(2*k, 2*k),
             (4*k, 4*k),(4*k, 2*k,2*k),]
        }
    ]

    X_res, y_res = sm.fit_resample(X, y)
    X_train, X_test, y_train, y_test = train_test_split(X_res, y_res, test_size=0.332,
                                                        stratify=y_res, random_state=42)

    # Configurar el clasificador con diferentes tamaños de capa oculta
    clf = GridSearchCV(pipeline, param_grid=param_grid, cv=3, scoring='accuracy')

    # Configurar el clasificador con diferentes tamaños de capa oculta

    clf.fit(X_train, y_train)

    test_pred = clf.predict(X_test)
    accuracy = accuracy_score(y_test, test_pred)

    results.append((file, clf.best_params_, accuracy))

# Ordenar los resultados por accuracy de mayor a menor
results.sort(key=lambda x: x[2], reverse=True)

# Mostrar el mejor resultado
best_file,best_clasi, best_accuracy = results[0]
print(f"Mejor archivo: {best_file}, Mejor accuracy: {best_accuracy:.2f},Mejor Modelo: {best_clasi}")

results=pd.DataFrame(results)
results.to_csv('/content/drive/MyDrive/Codigo Tesis Helicobacter Editado/Resultados/MLPClassifier.csv',index=False,sep=';',header=True )
